+++ packages/toolpad-app/src/components/Release.tsx
@@ -4,9 +4,9 @@ import * as React from 'react';
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import RocketLaunchIcon from '@mui/icons-material/RocketLaunch';
 import { useParams } from 'react-router-dom';
+import { NodeId } from '@mui/toolpad-core';
 import client from '../api';
 import * as appDom from '../appDom';
-import { NodeId } from '../types';
 import ToolpadAppShell from './ToolpadAppShell';
 
 interface NavigateToReleaseActionProps {
