+++ packages/toolpad-core/package.json
@@ -12,7 +12,8 @@
     "./package.json": "./package.json",
     ".": "./dist/index.js",
     "./constants": "./dist/constants.js",
-    "./runtime": "./dist/runtime.js"
+    "./runtime": "./dist/runtime.js",
+    "./utils": "./dist/utils.js"
   },
   "files": [
     "dist"
