+++ packages/toolpad-core/src/types.ts
@@ -1,5 +1,8 @@
-import React from 'react';
-import { TOOLPAD_COMPONENT } from './constants';
+import type * as React from 'react';
+import type { TOOLPAD_COMPONENT } from './constants';
+import type { Branded } from './utils';
+
+export type NodeId = Branded<string, 'NodeId'>;
 
 export type BindingAttrValueFormat = 'stringLiteral' | 'default';
